refactor: abstract InputView and OutputView
