+++ src/main/java/lotto/controller/LottoController.java
@@ -1,32 +1,33 @@
 package lotto.controller;
 
-import lotto.domain.dto.StatisticsResultDTO;
 import lotto.domain.dto.PurchasePriceInputDTO;
 import lotto.domain.dto.PurchaseResultDTO;
+import lotto.domain.dto.StatisticsResultDTO;
 import lotto.domain.dto.WinningLottoInputDTO;
 import lotto.service.LottoService;
 import lotto.view.View;
 
 public class LottoController {
 
-    private static final View view = new View();
-    private static final LottoService lottoService = new LottoService();
+    private final View view;
+    private final LottoService lottoService;
 
-    public static void main(String[] args) {
+    public LottoController(View view) {
+        this.view = view;
+        this.lottoService = new LottoService();
+    }
+
+    public void start() throws IllegalArgumentException {
         try {
-            start();
+            PurchasePriceInputDTO purchasePriceInputDTO = view.getPurchaseCost();
+            PurchaseResultDTO purchaseResultDTO = lottoService.purchase(purchasePriceInputDTO);
+            view.printLottoPurchaseResult(purchaseResultDTO);
+
+            WinningLottoInputDTO winningLottoInputDTO = view.getWinningLottoAndBonus();
+            StatisticsResultDTO statisticsResultDTO = lottoService.calculateResult(purchaseResultDTO, winningLottoInputDTO);
+            view.printLottoStatistics(statisticsResultDTO);
         } catch (IllegalArgumentException e) {
             view.printException(e.getMessage());
         }
-    }
-
-    public static void start() throws IllegalArgumentException {
-        PurchasePriceInputDTO purchasePriceInputDTO = view.getPurchaseCost();
-        PurchaseResultDTO purchaseResultDTO = lottoService.purchase(purchasePriceInputDTO);
-        view.printLottoPurchaseResult(purchaseResultDTO);
-
-        WinningLottoInputDTO winningLottoInputDTO = view.getWinningLottoAndBonus();
-        StatisticsResultDTO statisticsResultDTO = lottoService.calculateResult(purchaseResultDTO, winningLottoInputDTO);
-        view.printLottoStatistics(statisticsResultDTO);
     }
 }
